Összefésülési hibák javítása
