+++ bin/create_user
@@ -2,10 +2,14 @@
 <?php
 // create_user.php <name> <email> <password>
 use Webshop\Model\User;
-use Webshop\Core\EntityManagerFactory;
+use Webshop\EntityManagerFactory;
 
 require_once __DIR__ . '/../vendor/autoload.php';
-require __DIR__ . '/../src/Bootstrap.php';
+// .env konfigurációs fájl használata ($_ENV superglobalban)
+$dotenv = \Dotenv\Dotenv::createImmutable(__DIR__ . '/../');
+$dotenv->safeLoad();
+// EntityManager létrehozása
+$entityManager = EntityManagerFactory::getEntityManager();
 
 // Parancssori argumentumok feldolgozása
 $options = getopt('', ['role:']);
